+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+- Add HTTP endpoint to revoke all tokens for a user.
+
 ## v0.2.2 - 2025-04-14
 
 - Set initial focus to `sub` input in authorization form.